Conflicts:
	build.sh

+++ build.sh
@@ -14,7 +14,12 @@ cat css/bootstrap-tour.min.css css/dge.css css/venn.css css/slick.grid.css | cle
 
 echo "Compiling CoffeeScript and bundling all js..."
 browserify -t coffeeify -t hbsfy app/main.coffee > build/main.big.js 
-echo "Minifying js..."
-uglifyjs build/main.big.js > build/main.js
-rm build/main.big.js
 
+if [ "$1" = "local" ]; then
+  echo "Using unminified js"
+  mv  build/main.big.js build/main.js
+else
+  echo "Minifying js"
+  uglifyjs build/main.big.js > build/main.js
+  rm build/main.big.js
+fi
